# swagger.yaml
swagger: '2.0'
info:
  description: the description of apiinfo.
  version: 1.0.0
  title: Learn how to use swagger.
  contact:
    name: openjaw team
    email: heting.zhao@www.com
host: 'localhost:8080'
basePath: /
schemes:
  - http
  - https
tags:
  - name: Rule Management
    description: Download Zip
paths:
  /ruleBundles:
    post:
      tags:
        - Rule Management
      summary: Accept rule pushes request from console to store it.
      description: Accept rule pushes request from console to store it.
      operationId: pushRuleBundles
      produces:
        - application/json
      consumes:
        - multipart/form-data
      parameters:
        - name: types
          in: path
          description: rule types
          required: true
          type: string
        - name: tenantId
          in: path
          description: tenant id
          required: true
          type: string
        - name: version
          in: path
          description: rule version
          required: true
          type: string
        - name: ruleZip
          in: formData
          description: rule bundles
          required: true
          type: file
      responses:
        '200':
          description: RuleBundle ID
          schema:
            type: string
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
  /downLoadZipInFile:
    get:
      tags:
        - Rule Management
      summary: Load rules from SDK.
      description: >-
        Return a zip file as a byte array in customised response entity,then the
        file can be downloaded from front end.
      produces:
        - application/octet-stream
      consumes:
        - application/octet-stream
      responses:
        '200':
          description: Return a zip file.
          schema:
            type: file
        '400':
          description: Invalid ID supplied
        '404':
          description: not found
  /rules:
    get:
      tags:
        - Rule Management
      summary: Load rules from SDK.
      description: >-
        Return a zip file as a byte array in customised response entity,then the
        client can get the array.
      parameters:
        - name: type
          in: path
          description: rule types
          required: true
          type: string
        - name: ruleSetName
          in: path
          description: if business rule
          required: true
          type: string
        - name: tableName
          in: path
          description: 'if table rule,table set/table name'
          required: true
          type: string
        - name: tenantId
          in: path
          description: Hu B2c
          required: true
          type: string
        - name: version
          in: path
          description: v0.1.0
          required: true
          type: string
      responses:
        '200':
          description: >-
            Return a zip file as a byte array in customised response entity
            successfully.
          schema:
            $ref: '#/definitions/ruleResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: not found
  '/rules/subscribers/{subscriberId}/activatedInfo':
    post:
      tags:
        - RuleManagement
      summary: Report subscriber's latest information from sdk.
      description: fff.
      parameters:
        - name: subscriberId
          required: true
          in: path
          type: string
        - name: activatedRuleInfoList
          description: Activated Rule Info List.
          in: body
          schema:
            type: array
            items:
              $ref: '#/definitions/ActivatedInfo'
      responses:
        '200':
          description: Return a zip file.
          schema:
            type: string
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
definitions:
  ruleResponse:
    description: transform the result
    type: object
    properties:
      ruleInBytes:
        type: string
        format: byte
        description: put zip file in a byte array.
  ActivatedInfo:
    type: object
    description: jkljk
    properties:
      ruleGroup:
        type: string
        description: jkljk
      ruleList:
        description: jkljkarray
        type: array
        items:
          type: string
      version:
        type: string
        description: jkljk
      activatedTime:
        type: string
        description: current time